+++ frameworks/Java/greenlightning/greenlightning.dockerfile
@@ -1,11 +1,11 @@
-FROM maven:3.5.4-jdk-10-slim as maven
+FROM maven:3.5.4-jdk-8-slim as maven
 
 WORKDIR /greenlightning    
 COPY pom.xml pom.xml
 COPY src src
 RUN mvn clean install -q
 
-FROM openjdk:10-jdk-slim
+FROM openjdk:11-jdk-slim
 WORKDIR /greenlightning
 COPY --from=maven /greenlightning/target/greenlightning-test.jar app.jar
-CMD ["java", "-server", "-Xmx16g", "-XX:+UseNUMA", "-XX:+UseParallelGC", "-XX:+AggressiveOpts", "-jar", "app.jar"]
+CMD ["java", "-server", "-Xmx14g", "-XX:+UseNUMA", "-XX:+UseParallelGC", "-XX:+AggressiveOpts", "-jar", "app.jar"]
