+++ frameworks/Java/greenlightning/src/main/java/com/ociweb/gl/benchmark/GreenLightning.java
@@ -5,7 +5,10 @@ import com.ociweb.gl.api.GreenRuntime;
 public class GreenLightning {
 
 	public static void main(String[] args) {
-
+		
+		//PipeConfig.showConfigsCreatedLargerThan = 1<<23;
+		
+		//System.setProperty("pronghorn.processors", "28"); //TODO: could also lower memory usage by shrinking the stack space...
 		GreenRuntime.run(new FrameworkTest(),args);
 	
 	}
